Initial refactoring (not tested)
Fix build error

Update Echo engine to use toolkit

Fix async stream issue

Fix test: Add ability to specify CorpusService mock
