+++ src/Echo/src/EchoTranslationEngine/Usings.cs
@@ -5,3 +5,4 @@ global using Google.Protobuf.WellKnownTypes;
 global using Grpc.Core;
 global using Microsoft.Extensions.Diagnostics.HealthChecks;
 global using Serval.Translation.V1;
+global using SIL.ServiceToolkit.Utils;
