+++ src/Machine/src/Serval.Machine.Shared/Usings.cs
@@ -54,7 +54,7 @@ global using SIL.Machine.Tokenization;
 global using SIL.Machine.Translation;
 global using SIL.Machine.Translation.Thot;
 global using SIL.Machine.Utils;
-global using SIL.Scripture;
+global using SIL.ServiceToolkit.Models;
 global using SIL.ServiceToolkit.Services;
 global using SIL.ServiceToolkit.Utils;
 global using SIL.WritingSystems;
