+++ src/Machine/test/Serval.Machine.Shared.Tests/Services/SmtTransferEngineServiceTests.cs
@@ -687,7 +687,6 @@ public class SmtTransferEngineServiceTests
                         Substitute.For<ILogger<PreprocessBuildJob>>(),
                         _env.BuildJobService,
                         _env.SharedFileService,
-                        Substitute.For<ICorpusService>(),
                         _env._lockFactory,
                         _env.TrainSegmentPairs
                     )
