+++ src/ServiceToolkit/src/SIL.ServiceToolkit/Models/Row.cs
@@ -1,0 +1,3 @@
+namespace SIL.ServiceToolkit.Models;
+
+public record Row(string TextId, IReadOnlyList<object> Refs, string SourceSegment, string TargetSegment, int RowCount);
