+++ src/ServiceToolkit/src/SIL.ServiceToolkit/Services/CorpusService.cs
@@ -1,4 +1,4 @@
-﻿namespace Serval.Machine.Shared.Services;
+﻿namespace SIL.ServiceToolkit.Services;
 
 public class CorpusService : ICorpusService
 {
