+++ src/ServiceToolkit/src/SIL.ServiceToolkit/Services/ICorpusService.cs
@@ -1,4 +1,4 @@
-﻿namespace Serval.Machine.Shared.Services;
+﻿namespace SIL.ServiceToolkit.Services;
 
 public interface ICorpusService
 {
